requirements.txt
# Flask and extensions
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
WTForms==3.1.1

# Database
SQLAlchemy==2.0.23
alembic==1.13.0

# External API integration
requests==2.31.0

# Utility libraries
python-dotenv==1.0.0
Werkzeug==2.3.7

# Date and time utilities
python-dateutil==2.8.2

# Development tools
pytest==7.4.3
